SosADL2IOSTSGenerator: Do not generate concatenation transition between two statements when not needed. Previously, concatenation transitions were always generated between two statements. Now, a concatenation transition is generated only when the first statement ends with multiple states.
Example 1: 'if (cond) then { action1 } else { action2 }' ends with 2 states, the ending state of action1 and the ending state of action2.
If this if-then-else statement is followed by another statement, like in 'if (cond) then { action1 } else { action2 } action3',
we have to generate 2 concatenations transitions:
- one between the ending state of action1 and the first state of action3,
- one between the ending state of action2 and the first state of action3.

Example2: 'action1 action2': action1 ends with a single state,
thus there is no need to generate a concatenation transition between the ending state of action1 and the first state of action2.