SosADLPrettyPrinterGenerator.xtend: Changed the output of the roundtrip result, to make OK and KO results more distinguishable.
If the result of the roundtrip is ok (true/true/true), then the result is printed with 'ok' keyword in stdout.
If the result of the roundtrip is not ok, then the result is printed with 'KO' keyword in stderr.
In the eclipse output, messages appear in black if displayed in stdout, and in red if displayed in stderr.