SosADL2IOSTSGenerator.xtend: bug fix when generating a Valuing out of DoExpr.
Recall that if a DoExpr occurs in a behavior in SosADL, it is transformed into a Valuing, which type is the type of the doExpr.
Since we don't have a type checker at this time, the Valuing is "value _doExprResult# is TYPE_TODO# = doExpr",
where:
- _doExprResult# is a new variable name, and # is a new index,
- TYPE_TODO# is a NamedType, and # is the same index,
- doExpr is the DoExpr expression.
Previously1: the TYPE_TODO was not indexed, like the variable _doExprResult#: fixed.
Previously2: the TYPE_TODO was replaced by an integer, since we assumed an undeclared type was an 'integer'.
This assumption was false in this case, since the type named "TYPE_TODO#" means we shall compute it!