Changes to migrate from Xtext 2.8 (Eclipse Mars.1) to Xtext 2.9 (Eclipse Mars.1).
These are the minimal changes to apply in org.archware.sosadl project:
- new workflow: src/org/archware/sosadl/GenerateSosADL.mwe2
- new generator: src/org/archware/sosadl/generator/SosADLGenerator.xtend
- these java classes are removed because they are translated in xtend:
  src/org/archware/sosadl/SosADLRuntimeModule.java
  src/org/archware/sosadl/SosADLStandaloneSetup.java

First git pull this commit to apply these changes to your git repository.
Then, clean your workspace, and proceed to the installation of the org.archware.sosadl
as a new Xtext project in Eclipse Mars.2. This will generate all other files.