SosADL2IOSTSGenerator.xtend: fixed the connection name in actions inside gate's or duty's protocols.
The type-checker does not allow to prefix the connection name with the gate's or duty's name within actions inside gate or duty protocols.
For example, before the type-checker, we could write:
  gate gate1 is { connection connection1 is in{type1} }
    guarantee {
    protocol gate1protocol is {
      via gate1::connection1 receive any
      done }}
Now, we have to omit the prefix "gate1::" in the action, thus we write in "via connection1 receive any".

The fix consists of retreiving the gate or duty's name, in order to generate the IoSTS as before.
- One method has been added: def String getNameOfGateOrDuty(ProtocolAction a)
  which travels the containers of the ProtocolAction until it gets the GateDecl or DutyDecl.
- The method computeSTS(int startState, ProtocolAction a) now uses the above method
  to complete the connection'name in the ProtocolAction with the gate's or duty's name.