SosADL2IOSTSGenerator.xtend: the SosADL2IOSTSGenerator now recognises boolean expressions as such.
Added: def dispatch IOstsType computeIOstsType(BooleanType t).

Previously, the function IOstsType computeIOstsType(DataType t) didn't recognised type of t if it was a BooleanType.
Now, if t is an instance of BooleanType, a instance of IOstsBoolType is created.

This was an important bug, because boolean expression can occur in SoS-ADL even if the BooleanType is not a valid datatype in valuing.
Indeed, boolean expressions are allowed in "do expression", and in this case, one has to declare a variable of type boolean in IoSTS.